refactor: split client.py into domain modules via mixins
client.py was 1185 lines mixing unrelated domains. Split into focused
mixin modules; SuperMeClient inherits all. Zero functional change.
