+++ superme_sdk/models.py
@@ -1,0 +1,47 @@
+"""Response model classes (mirror OpenAI SDK objects)."""
+
+from __future__ import annotations
+
+from typing import Optional
+
+
+class Message:
+    """A single chat message."""
+
+    def __init__(self, data: dict) -> None:
+        self.role: str = data.get("role", "assistant")
+        self.content: str = data.get("content", "") or ""
+
+
+class Choice:
+    """One completion choice."""
+
+    def __init__(self, data: dict) -> None:
+        self.index: int = data.get("index", 0)
+        self.message = Message(data.get("message", {}))
+        self.finish_reason: Optional[str] = data.get("finish_reason")
+
+
+class Usage:
+    """Token usage statistics."""
+
+    def __init__(self, data: dict) -> None:
+        self.prompt_tokens: int = data.get("prompt_tokens", 0)
+        self.completion_tokens: int = data.get("completion_tokens", 0)
+        self.total_tokens: int = data.get("total_tokens", 0)
+
+
+class ChatCompletion:
+    """OpenAI-compatible chat completion response.
+
+    SuperMe-specific fields (``metadata``) are preserved as attributes.
+    """
+
+    def __init__(self, data: dict) -> None:
+        self.id: str = data.get("id", "")
+        self.object: str = data.get("object", "chat.completion")
+        self.created: int = data.get("created", 0)
+        self.model: str = data.get("model", "")
+        self.choices: list[Choice] = [Choice(c) for c in data.get("choices", [])]
+        self.usage = Usage(data.get("usage") or {})
+        self.metadata: Optional[dict] = data.get("metadata")
